Change config/setup_node.py shebang
The config/setup_node.py script would not work correctly if invoked
directly when in a virtualenv. This was because the shebang was set to
"/usr/bin/python" and required packages weren't available in the
system python package search path.

The shebang is now set to "/usr/bin/env python" which uses the
virtualenv python interpreter and accompanying package search
path. The config/setup_node.py now works correctly when directly
invoked since it can find the required packages which are installed in
the active virtualenv.