Add dependencies to setup.py, update docs, script.
The setup.py file now contains all the dependencies for the LR Pylons
app in the 'install_requires' list. Additionally, the setup_requires
entry was removed because Paster is installed as a Pylons dependency
and Pylons is declared in the install_requires list. Due to this
change, using 'pip install -e .' installs all the declared
dependencies.

Modified the README.md file and the install_pydeps.bash script to
reflect the changes to the setup.py file. Setup is even simpler now.