sudo and Xcode 4 related fixes for MacOS.
Install pip and virtualenv using sudo if /usr/local aren't writable by
the user running the install script. Similarly, use use sudo when
installing couchdb using Homebrew if it is installed such that sudo is
required when using brew.

Don't install lxml if Xcode 4 is installed on MacOS. The script prints
out links to manual installation instructions. lxml is no longer a
requirement in the setup.py script. Instead it is separately by the
setup script unless Xcode 4 is detected.

Updated the README.md file to indicate that there are issues with
building lxml with Xcode 4.