updated JMeter Test to do some fun stuff..
There is a known issue that I'm trying to determine if it's a bug with the test or an actual bug in code.  Occasionally, the OAI-PMH GetRecord fails because the wrong metadataPrefix is used in the request.  Unknown if the ListIdentifiers is returning a bad set or if some variable isn't thread safe in JMeter.

Signed-off-by: Jim Klo <jim.klo@sri.com>