Create service template for all services
Create service template for all services to fix #125 and issue #34.
This change makes it possible to active/deactive all services during setup.

Add setup option to specify if the node is gateway node or not. When
the node is a gateway node setup only the services that allowed to
run on a gateway node.