Update Distribute To Closer Match Spec
First pass at updating the distribute code to match the spec closer.  It is
not in full compliance because there are some required fields/data that are not
mentioned in the spec. 'filter_description' and 'resource_data_url' are additional
fields that a source node needs from the destination node complete
distribute/replication