Allow distribution gateway to common node on the same network
Update the distribute code to allow gateway nodes as source node that distributes
to common nodes that are in the same network and community.

Update the tests to reflect that distribute is allow between gateway and common
nodes when they are in the same network and community.

Fix some bugs in large_network_test and in the lr_node comparison