Fix issue where when no term is being specified, one was being generated plus adding a generated tstamp, causing over filtering to occur
Problem mainly occurs with timestamp.  If no term (resource or discriminator depending upon the view) was specified for a view that includes a TS, the service would end up generating a key for it, which would cause the view to often return nothing because of a condition where [ "", min-TS] to [ "\ud7af", max-TS ] since the first element has not valid range.