try making test wait until indexer is idle before trying to start
i think what's happening is the indexer is still active when starting the test, so publishing new data makes it get indexed faster.

This should cause the indexer to wait up to 30 seconds for it to go idle… if this doesn't work we'll need to probably either need to just skip this test (boo) or make the wait time longer.